pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "Retrieval-Augmented Generation chatbot for semantic search and Q&A on PDF documents"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {text = "MIT"}
keywords = ["rag", "langchain", "streamlit", "llm", "semantic-search", "chatbot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "langchain>=1.2.15",
    "langchain-core>=1.2.28",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.1.1",
    "langchain-groq>=1.1.2",
    "sentence-transformers>=5.3.0",
    "chromadb>=1.5.7",
    "python-dotenv>=1.0.0",
    "streamlit>=1.28.0",
    "onnxruntime>=1.18.0,<1.24.0",
]
